requirements.txt
pandas
scikit-learn
streamlit_extras
streamlit_option_menu
streamlit_elements
streamlit_echarts
xgboost
joblib